The default queue "driver" is actually a "connection"
The description here has bothered me for a while, because it is kind of misleading.

The `QUEUE_DRIVER` environment variable perhaps also needs changing to `QUEUE_CONNECTION`, but I'm not sure if that is just too engrained in legacy systems now? I can change that on this PR if you agree, and also the matching `QUEUE_DRIVER=sync` in `.env.example`.